bookdel
This GitHub repository houses the code for a fully functional e-commerce website specializing in books.  It's designed to provide a comprehensive platform for users to browse, discover, purchase, and manage their book collections online.  This project serves as a valuable learning resource for developers interested in building e-commerce applications, particularly within the niche of book sales.  It showcases best practices in web development, database management, and secure transaction processing.  The repository includes front and backend components, allowing developers to set up and run the website locally.